Use fast-forward merging.
Initially I thought it was a good idea to use --no-ff, but since we are
oftenly dealing with single commit pull requests this has to go. It's just
annoying to have. If you still want it, just manually add that option to
the call. However, it's no longer the default.