Use OAuth to authenticate
Do not store username/password in plain text inside of ~/gitconfig.
Request a OAuth token and store it inside of a dedicated
file called ~/.git_review.yml.

Using OAuth is safer because GitHub's password is no longer readable
by other users of the system, plus access to git-review can be revoked
using GitHub's applications page.

I also like to store the token (which is still a sensitive information)
outside of ~/.gitconfig. I keep this file inside of a public git
repository and I do not want to expose this private data to everybody.