Remove unnecessary check for Request class
Back in the time, the request variable in this method could have been
something else when the Github provider didn't use the Request model.

Now this is always the case, so this check is unnecessary. In addition
the repo.head.user.login is cleaner to use than
request.head.repo.owner.login because the repo owner might get nil.